Separate redirects logic from legacy redirect code
